% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foodbanksHamilton.R
\docType{data}
\name{foodbank_table}
\alias{foodbank_table}
\title{Description of foodbanks in Hamilton.}
\format{
A simple features data frame with 5 rows and 5 variables:
\describe{
  \item{Type}{Type of facility}
  \item{Description}{Address of facility}
  \item{Locations Pre-COVID}{Status of facility}
  \item{Locations During COVID}{Address of facility}
  \item{Notes}{Explanatory notes}
}
}
\usage{
data(foodbank_table)
}
\description{
Description of foodbanks and related meal services in the City of Hamilton.
}
\keyword{datasets}
\keyword{land}
\keyword{transportation}
\keyword{use}
