% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foodbanksHamilton.R
\docType{data}
\name{hamilton_boundary}
\alias{hamilton_boundary}
\title{Hamilton CMA.}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1 rows and 3 columns.
}
\source{
Statistics Canada
}
\usage{
data(hamilton_boundary)
}
\description{
Boundary of the City of Hamilton, in Ontario, Canada.
}
\keyword{Canada}
\keyword{Census}
\keyword{datasets}
