% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foodbanksHamilton.R
\docType{data}
\name{modes_less40k}
\alias{modes_less40k}
\title{Transportation for low income population by Traffic Analysis Zone.}
\format{
A simple features data frame with 234 rows and 5 variables:
\describe{
  \item{TAZUID}{Unique identifier of traffic analysis zone, according to the GTA06 zoning system of the Transportation Tomorrow Survey}
  \item{Transit}{Number of trips by Transit (excluding GO rail)}
  \item{Walk}{Number of trips by walking}
  \item{Driver}{Number of trips by Auto (driver)}
  \item{geometry}{Geometry of points}
}
}
\source{
Transportation Tomorrow Survey (http://dmg.utoronto.ca/transportation-tomorrow-survey/tts-introduction)
}
\usage{
data(modes_less40k)
}
\description{
Modes of transportation by traffic analysis zones for population in households with income less than $40,000. The source data is a cross-tabulation of primary mode of travel by household income by traffic analysis zone of the household.
}
\keyword{datasets}
\keyword{land}
\keyword{transportation}
\keyword{use}
