% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foodbanksHamilton.R
\docType{data}
\name{ttm_walk}
\alias{ttm_walk}
\title{Travel Time Matrix Walking}
\format{
A data frame with 4,693,844 rows and 3 variables:
\describe{
  \item{UID}{Unique identifier of parcel}
  \item{OBJECTID}{Unique identifier of pharmacy}
  \item{travel_time}{Travel time by transit from parcel to pharmacy in minutes}
}
}
\usage{
data(ttm_walk)
}
\description{
Travel time matrix by walking to pharmacies from nearest points on the network corresponding to residential parcels in Hamilton. This matrix was calculated using package \code{r5r} and a cutoff time of 180 minutes.
}
\keyword{datasets}
\keyword{transportation}
